refactor: extract text editor CodeMirror lifecycle
